Further improvement on movement of PlayerShip. Now it works more like the way I intend it to: you cannot move freely in all directions, but you can travel the direction the ship is currently facing, and turn using other keys. Some minor changes needed to be done - for instance, _x and _y in MapObject need to be doubles instead of ints.

+++ include/mapobjects/MapObject.h
@@ -25,10 +25,10 @@ class MapObject {
          */
         void draw(SDL_Renderer* renderer, int x, int y, SDL_Point* center = nullptr,
                   SDL_RendererFlip flip = SDL_FLIP_NONE);
-        int getX();
-        int getY();
+        double getX();
+        double getY();
     protected:
-        int _x, _y;
+        double _x, _y;
 
         /**
          * The angle (number of degrees) that the MapObject's sprite will be rotated to
